Relocate ServiceInfo to zeroconf._services.info
- This is an internal relocation only
