+++ tests/test_aio.py
@@ -16,7 +16,8 @@ from zeroconf.aio import AsyncServiceInfo, AsyncServiceListener, AsyncZeroconf, 
 from zeroconf import Zeroconf
 from zeroconf.const import _LISTENER_TIME
 from zeroconf._exceptions import BadTypeInNameException, NonUniqueNameException, ServiceNameAlreadyRegistered
-from zeroconf._services import ServiceInfo, ServiceListener
+from zeroconf._services import ServiceListener
+from zeroconf._services.info import ServiceInfo
 from zeroconf._utils.time import current_time_millis
 
 from . import _clear_cache
