+++ tests/test_services.py
@@ -19,11 +19,8 @@ import zeroconf as r
 from zeroconf import DNSAddress, DNSPointer, DNSQuestion, const, current_time_millis
 import zeroconf._services as s
 from zeroconf import Zeroconf
-from zeroconf._services import (
-    ServiceBrowser,
-    ServiceInfo,
-    ServiceStateChange,
-)
+from zeroconf._services import ServiceBrowser, ServiceStateChange
+from zeroconf._services.info import ServiceInfo
 from zeroconf.aio import AsyncZeroconf
 
 from . import has_working_ipv6, _clear_cache, _inject_response
